Move a lot of the atlas logic to Common/Render/AtlasGen.cpp/h, code cleanup
